+++ packages/falcon-server-env/src/models/Extension.ts
@@ -1,27 +1,34 @@
 import { GraphQLResolveInfo } from 'graphql';
+import { EventEmitter2 } from 'eventemitter2';
 import ApiDataSource from './ApiDataSource';
-import { ApiUrlPriority, ConfigurableConstructorParams, FetchUrlResult, ExtensionContainer } from '../types';
+import { ApiUrlPriority, ConfigurableConstructorParams, ExtensionContainer, FetchUrlResult } from '../types';
+
+export type ConfigurableContainerConstructorParams = ConfigurableConstructorParams & {
+  extensionContainer: ExtensionContainer;
+};
 
 export default abstract class Extension<TApiConfig = object> {
   public config: object;
   public name: string;
   public api?: ApiDataSource;
   public apiConfig: TApiConfig | null = null;
-  private extensionContainer: ExtensionContainer;
+
+  protected extensionContainer: ExtensionContainer;
+  protected eventEmitter: EventEmitter2;
 
   /**
-   * @param {object} config Extension config object
-   * @param {string} name Extension short-name
-   * @param {ExtensionContainer} extensionContainer Instance of ExtensionContainer
+   * @param {ConfigurableContainerConstructorParams} params Constructor params
+   * @param {object} params.config Extension config object
+   * @param {string} params.name Extension short-name
+   * @param {ExtensionContainer} params.extensionContainer ExtensionContainer instance
+   * @param {EventEmitter2} params.eventEmitter EventEmitter2 instance
    */
-  constructor({
-    config = {},
-    name,
-    extensionContainer
-  }: ConfigurableConstructorParams & { extensionContainer: ExtensionContainer }) {
+  constructor(params: ConfigurableContainerConstructorParams) {
+    const { config = {}, name, extensionContainer, eventEmitter } = params;
     this.name = name || this.constructor.name;
     this.config = config;
     this.extensionContainer = extensionContainer;
+    this.eventEmitter = eventEmitter;
   }
 
   /**